+++ unittests/Parse/BuildConfigTests.cpp
@@ -1,7 +1,8 @@
-#include "gtest/gtest.h"
-#include "llvm/ADT/Optional.h"
 #include "swift/Basic/SourceLoc.h"
 #include "swift/Basic/Version.h"
+#include "swift/Parse/ParseVersion.h"
+#include "llvm/ADT/Optional.h"
+#include "gtest/gtest.h"
 
 using namespace swift;
 using namespace llvm;
@@ -11,19 +12,16 @@ class VersionTest : public ::testing::Test{};
 class CompilerVersionUnpackingTest : public ::testing::Test {};
 
 Optional<version::Version> CV(const char *VersionString) {
-  return version::Version::parseCompilerVersionString(VersionString,
-                                                      SourceLoc(),
-                                                      nullptr);
+  return VersionParser::parseCompilerVersionString(VersionString, SourceLoc(),
+                                                   nullptr);
 }
 
 Optional<version::Version> V(const char *VersionString) {
-  return version::Version::parseVersionString(VersionString,
-                                              SourceLoc(),
-                                              nullptr);
+  return VersionParser::parseVersionString(VersionString, SourceLoc(), nullptr);
 }
 
 TEST_F(CompilerVersionTest, VersionComparison) {
-  auto currentVersion = version::Version::getCurrentCompilerVersion();
+  auto currentVersion = version::getCurrentCompilerVersion();
   EXPECT_GE(CV("700").getValue(), CV("602").getValue());
   EXPECT_GE(CV("700.*").getValue(), CV("700.*").getValue());
   EXPECT_GE(CV("700.*.1").getValue(), CV("700.*.0").getValue());
@@ -76,7 +74,7 @@ TEST_F(CompilerVersionUnpackingTest, VersionComparison) {
   // Since this test was added during 5.7, we expect all of these comparisons to
   // be GE, either because we are comparing to the empty version or because we
   // are comparing to a version >= 5.7.0.0.0.
-  auto currentVersion = version::Version::getCurrentCompilerVersion();
+  auto currentVersion = version::getCurrentCompilerVersion();
   EXPECT_GE(CV("700"), currentVersion);
   EXPECT_GE(CV("1300"), currentVersion);
   EXPECT_GE(CV("5007"), currentVersion);
